requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.9.9
djangorestframework>=3
djangorestframework-simplejwt>=5
django-cors-headers>=3.13.0
requests>=2.31.0
django-environ>=0.11.2
django-crontab>=0.7.1